fix test failing
